// src/webhook-dispatcher/domain/types/webhook-dispatcher-event.types.ts
import { NotificationStatus } from '../../../common/constants/notification-status.constants';

export interface WebhookEvent {
  notificationId: string;
  status: NotificationStatus;
  callbackUrl: string;
}

export interface WebhookCallbackPayload {
  notificationId: string;
  status: NotificationStatus;
}